FinalTermProject - Final Commit
the output "It is a Connect Four. X wins!" is updated to show the player who wins the game. To display the name of the winning player instead of the symbol 'X' :  the ConnectFourGame class is modified to public bool IsGameOver(out bool restart, out Player winningPlayer). and the GameController class is updated.